Added support for equipment
Enhanced code to allow for more object-oriented data access
- New EntityManager class with static helper functions for entities
- New architecture with <Entity>.php POPOs and <Entity>Model.php DAOs
- Using stored procedures instead of SQL prepared statements

+++ src/models/CategoryModel.php
@@ -4,11 +4,12 @@
 namespace app\models;
 
 use framework\Database;
+use framework\EntityManager;
 
 class CategoryModel
 {
 
-  private static $table = "categories_equipement";
+  public static $table = "categories_equipement";
 
   public static function getAll()
   {
@@ -24,10 +25,11 @@ class CategoryModel
       " ",
       [
         "select *",
-        "from " . Database::table(CategoryModel::$table),
+        "from " . Database::table(self::$table),
         "where parent is null or parent = ''"
       ]);
-    $rs = Database::getPDO()->query($sql);
+
+    $rs = Database::getPDO()->query("select * from {$_SESSION['dataset']['id']}_vu_category_getallmain");
     return $rs->fetchAll(\PDO::FETCH_ASSOC);
   }
 
@@ -42,13 +44,13 @@ class CategoryModel
         "c.libelle as libelle,",
         "p.code as code_parent,",
         "p.libelle as libelle_parent",
-        "from " . Database::table(CategoryModel::$table) . " as c",
-        "left join ". Database::table(CategoryModel::$table) . " as p",
+        "from " . Database::table(self::$table) . " as c",
+        "left join ". Database::table(self::$table) . " as p",
         "on c.parent = p.code",
         "order by c.parent, c.code"
       ]);
 
-    $rs = Database::getPDO()->query($sql);
+    $rs = Database::getPDO()->query("select * from {$_SESSION['dataset']['id']}_vu_category_getallwithmain");
     return $rs->fetchAll(\PDO::FETCH_ASSOC);
   }
 
@@ -65,9 +67,11 @@ class CategoryModel
         "from",
         Database::table(CategoryModel::$table) . " as c,",
         Database::table(CategoryModel::$table) . " as p",
-        "where c.parent is not null and c.parent = p.code"
+        "where c.parent is not null",
+        "and c.parent = p.code"
       ]);
-    $rs = Database::getPDO()->query($sql);
+
+    $rs = Database::getPDO()->query("select * from {$_SESSION['dataset']['id']}_vu_category_getallsubwithmain");
     return $rs->fetchAll(\PDO::FETCH_ASSOC);
   }
 
@@ -77,42 +81,64 @@ class CategoryModel
       Database::getOneQuery(self::$table, ["code"])
     );
     $statement->execute([$id]);
-    return $statement->fetch(\PDO::FETCH_ASSOC);
+    $data =  $statement->fetch(\PDO::FETCH_ASSOC);
+    return EntityManager::hydrate(Category::class, $data);
   }
 
   public static function insert($data)
   {
-    $statement = Database::getPDO()->prepare(
-      Database::insertQuery(
-        self::$table,
-        [ "code", "libelle", "parent" ]
-      )
-    );
-    return $statement->execute($data);
+    if ($data instanceof Category) {
+      $statement = Database::getPDO()->prepare("call {$_SESSION['dataset']['id']}_sp_category_insert(:code, :libelle, :parent);");
+      $statement->bindValue(':code', $data->getCode());
+      $statement->bindValue(':libelle', $data->getLibelle());
+      $statement->bindValue(':parent', $data->getParent());
+      return $statement->execute();
+    } else {
+      $statement = Database::getPDO()->prepare(
+        Database::insertQuery(
+          self::$table,
+          ["code", "libelle", "parent"]
+        )
+      );
+      return $statement->execute($data);
+    }
   }
 
   public static function update($data)
   {
-    $statement = Database::getPDO()->prepare(
-      Database::updateQuery(
-        self::$table,
-        [ "libelle", "parent" ],
-        [ "code" ]
-      )
-    );
-    return $statement->execute($data);
+    if ($data instanceof Category) {
+      $statement = Database::getPDO()->prepare("call {$_SESSION['dataset']['id']}_sp_category_update(:code, :libelle, :parent);");
+      $statement->bindValue(':code', $data->getCode());
+      $statement->bindValue(':libelle', $data->getLibelle());
+      $statement->bindValue(':parent', $data->getParent());
+      return $statement->execute();
+    } else {
+      $statement = Database::getPDO()->prepare(
+        Database::updateQuery(
+          self::$table,
+          ["libelle", "parent"],
+          ["code"]
+        )
+      );
+      return $statement->execute($data);
+    }
   }
 
-
-  public static function deleteOne($id)
+  public static function deleteOne($data)
   {
-    $statement = Database::getPDO()->prepare(
-      Database::deleteOneQuery(
-        self::$table,
-        [ "code" ]
-      )
-    );
-    return $statement->execute([$id]);
+    if ($data instanceof Category) {
+      $statement = Database::getPDO()->prepare("call {$_SESSION['dataset']['id']}_sp_category_delete(:code);");
+      $statement->bindValue(':code', $data->getCode());
+      return $statement->execute();
+    } else {
+      $statement = Database::getPDO()->prepare(
+        Database::deleteOneQuery(
+          self::$table,
+          ["code"]
+        )
+      );
+      return $statement->execute([ $data ]);
+    }
   }
 
   public static function getProperties($id)
@@ -121,6 +147,24 @@ class CategoryModel
       [
         Database::getAllQuery("categories_proprietes"),
         Database::buildWhere(["code_categorie"])
+      ]);
+    $statement = Database::getPDO()->prepare($sql);
+    $statement->execute([$id]);
+    return $statement->fetchAll(\PDO::FETCH_ASSOC);
+  }
+
+  public static function getPropertyList($id)
+  {
+    $sql = implode(" ",
+      [
+        "select",
+        "cp.code_propriete as code,",
+        "pr.intitule as intitule",
+        "from",
+        Database::table("categories_proprietes") . " as cp",
+        "join " . Database::table("proprietes_equipement") . " as pr",
+        "on pr.code = cp.code_propriete",
+        Database::buildWhere([ "cp.code_categorie" ])
       ]);
     $statement = Database::getPDO()->prepare($sql);
     $statement->execute([$id]);
